Move Game.currentBoard().createNextBoard() → game.nextBoard()
- `Board` should't deal with making copies of itself
- Cleaner syntax
